+++ lib/view_component/polymorphic_slots.rb
@@ -5,6 +5,15 @@ module ViewComponent
     # In older rails versions, using a concern isn't a good idea here because they appear to not work with
     # Module#prepend and class methods.
     def self.included(base)
+      if base != ViewComponent::Base
+        location = Kernel.caller_locations(1, 1)[0]
+
+        warn(
+          "warning: ViewComponent::PolymorphicSlots is now included in ViewComponent::Base by default "\
+          "and can be removed from #{location.path}:#{location.lineno}"
+        )
+      end
+
       base.singleton_class.prepend(ClassMethods)
       base.include(InstanceMethods)
     end
@@ -45,8 +54,12 @@ module ViewComponent
               "#{slot_name}_#{poly_type}"
             end
 
-          # Deprecated: Will be removed in 3.0
           define_method(setter_name) do |*args, &block|
+            ViewComponent::Deprecation.warn(
+              "polymorphic slot setters like `#{setter_name}` are deprecated and will be removed in"\
+              "ViewComponent v3.0.0.\n\nUse `with_#{setter_name}` instead."
+            )
+
             set_polymorphic_slot(slot_name, poly_type, *args, &block)
           end
           ruby2_keywords(setter_name.to_sym) if respond_to?(:ruby2_keywords, true)